+++ packages/field-error/examples/default.tsx
@@ -1,0 +1,7 @@
+import { FieldError, type FieldErrorProps } from '@bento/field-error';
+/* v8 ignore next */
+import React from 'react';
+
+export function DefaultExample(args: FieldErrorProps) {
+  return <FieldError {...args}>This field is required</FieldError>;
+}