+++ packages/visually-hidden/src/index.tsx
@@ -1,0 +1,35 @@
+import React from 'react';
+import { withSlots } from '@bento/slots';
+import { useProps } from '@bento/use-props';
+import { VisuallyHiddenProps as AriaVisuallyHiddenProps, useVisuallyHidden } from 'react-aria';
+import { useDataAttributes } from '@bento/use-data-attributes';
+
+export interface VisuallyHiddenProps extends AriaVisuallyHiddenProps {
+  /** The element type to render the component as. Community standard (preferred) @default 'span' */
+  as?: keyof JSX.IntrinsicElements;
+
+  /**
+   * @deprecated Use `as` instead.
+   * The element type to render the component as. For React Aria compatibility. @default 'span' */
+  elementType?: keyof JSX.IntrinsicElements;
+}
+
+/**
+ * The `VisuallyHidden` component is used to hide content visually while keeping it accessible to screen readers and assistive technologies.
+ */
+export const VisuallyHidden = withSlots('BentoVisuallyHidden', function VisuallyHidden(args: VisuallyHiddenProps) {
+  const { props, apply } = useProps(args);
+  const { as, elementType, ...rest } = props;
+  const Element = as ?? elementType ?? 'span';
+  const { visuallyHiddenProps } = useVisuallyHidden(rest);
+
+  return (
+    <Element
+      {...apply(visuallyHiddenProps, ['isFocusable'])}
+      {...useDataAttributes({
+        // If the style is set, the content is visually hidden
+        hidden: Boolean(visuallyHiddenProps.style)
+      })}
+    />
+  );
+});